1) Gemini AI Assistant Bot/requirements.txt
streamlit==1.42.2
python-dotenv==1.1.0
speechrecognition==3.14.1
google-generativeai==0.3.2
protobuf==3.20.3
sympy==1.13.3
wikipedia_api==0.8.1